requirements.txt
Django==1.9.5
django-bootstrap-form==3.2
django-bower==5.1.0
django-colorfield==0.1.10
django-polymorphic==0.9.1
https://github.com/xiscosc/django-sendsms/archive/developer.zip
psycopg2==2.6.1
pushbullet.py==0.10.0
python-magic==0.4.11
requests==2.9.1
six==1.10.0
websocket-client==0.35.0
django-async-messages
sendgrid-django